Factor the config path out into the lib.
